# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Photos API - Google Photos-like app backed by Azure Blob Storage
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: string;
  name: string;
  description?: string;
  coverUrl?: string;
  photoCount: number;
  createdAt: Date;
}