# lib/api-zod/src/generated/types/photoListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Photos API - Google Photos-like app backed by Azure Blob Storage
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface PhotoListResponse {
  photos: Photo[];
  nextCursor?: string;
  total: number;
}