# lib/api-zod/src/generated/types/photoStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Photos API - Google Photos-like app backed by Azure Blob Storage
 * OpenAPI spec version: 0.1.0
 */

export interface PhotoStats {
  total: number;
  favorites: number;
  trashed: number;
  albums: number;
  totalSize: number;
}